+++ graphicsbuilder/trunk/graphicsbuilder/src/main/groovy/groovy/swing/j2d/operations/shapes/CrossGraphicsOperation.groovy
@@ -60,11 +60,11 @@ public class CrossGraphicsOperation extends AbstractShapeGraphicsOperation {
        def w = width != null ? width : radius * 3 / 5
        def a = angle != null ? angle : 0
        def n = roundness != null ? roundness : 0
-       cross = new Cross( cx as double,
-                          cy as double,
-                          radius as double,
-                          w as double,
-                          a as double,
-                          n as double )
+       cross = new Cross( cx as float,
+                          cy as float,
+                          radius as float,
+                          w as float,
+                          a as float,
+                          n as float )
     }
 }
