+++ graphicsbuilder/trunk/graphicsbuilder/src/main/java/groovy/swing/j2d/geom/MultiRoundRectangle.java
@@ -28,30 +28,30 @@ import java.awt.geom.Rectangle2D;
  * @author Andres Almiray <aalmiray@users.sourceforge.net>
  */
 public class MultiRoundRectangle implements Shape, Cloneable {
-   private double bottomLeftHeight;
-   private double bottomLeftWidth;
-   private double bottomRightHeight;
-   private double bottomRightWidth;
-   private double height;
+   private float bottomLeftHeight;
+   private float bottomLeftWidth;
+   private float bottomRightHeight;
+   private float bottomRightWidth;
+   private float height;
    private GeneralPath rectangle;
-   private double topLeftHeight;
-   private double topLeftWidth;
-   private double topRightHeight;
-   private double topRightWidth;
-   private double width;
-   private double x;
-   private double y;
+   private float topLeftHeight;
+   private float topLeftWidth;
+   private float topRightHeight;
+   private float topRightWidth;
+   private float width;
+   private float x;
+   private float y;
 
-   public MultiRoundRectangle( double x, double y, double width, double height, double topLeft,
-         double topRight, double bottomLeft, double bottomRight ) {
+   public MultiRoundRectangle( float x, float y, float width, float height, float topLeft,
+         float topRight, float bottomLeft, float bottomRight ) {
       this( x, y, width, height, topLeft, topLeft, topRight, topRight, bottomLeft, bottomLeft,
             bottomRight, bottomRight );
    }
 
-   public MultiRoundRectangle( double x, double y, double width, double height,
-         double topLeftWidth, double topLeftHeight, double topRightWidth, double topRightHeight,
-         double bottomLeftWidth, double bottomLeftHeight, double bottomRightWidth,
-         double bottomRightHeight ) {
+   public MultiRoundRectangle( float x, float y, float width, float height,
+         float topLeftWidth, float topLeftHeight, float topRightWidth, float topRightHeight,
+         float bottomLeftWidth, float bottomLeftHeight, float bottomRightWidth,
+         float bottomRightHeight ) {
       if( topLeftWidth + topRightWidth > width ){
          throw new IllegalArgumentException( "top rounding factors are invalid (" + topLeftWidth
                + ") (" + topRightWidth + ")" );
@@ -106,11 +106,11 @@ public class MultiRoundRectangle implements Shape, Cloneable {
       return rectangle.contains( r );
    }
 
-   public double getBottomLeft() {
+   public float getBottomLeft() {
       return bottomLeftWidth;
    }
 
-   public double getBottomRight() {
+   public float getBottomRight() {
       return bottomRightWidth;
    }
 
@@ -122,7 +122,7 @@ public class MultiRoundRectangle implements Shape, Cloneable {
       return rectangle.getBounds2D();
    }
 
-   public double getHeight() {
+   public float getHeight() {
       return height;
    }
 
@@ -134,15 +134,15 @@ public class MultiRoundRectangle implements Shape, Cloneable {
       return rectangle.getPathIterator( at, flatness );
    }
 
-   public double getTopLeft() {
+   public float getTopLeft() {
       return topLeftWidth;
    }
 
-   public double getTopRight() {
+   public float getTopRight() {
       return topRightWidth;
    }
 
-   public double getWidth() {
+   public float getWidth() {
       return width;
    }
 
@@ -158,7 +158,7 @@ public class MultiRoundRectangle implements Shape, Cloneable {
       rectangle = new GeneralPath();
       if( topLeftWidth > 0 ){
          rectangle.moveTo( x + topLeftWidth, y );
-         rectangle.append( new Arc2D.Double( x, y, topLeftWidth * 2, topLeftHeight * 2, 90, 90,
+         rectangle.append( new Arc2D.Float( x, y, topLeftWidth * 2, topLeftHeight * 2, 90, 90,
                Arc2D.OPEN ), true );
       }else{
          rectangle.moveTo( x, y );
@@ -166,14 +166,14 @@ public class MultiRoundRectangle implements Shape, Cloneable {
       }
 
       if( bottomLeftWidth > 0 ){
-         rectangle.append( new Arc2D.Double( x, y + height - (bottomLeftHeight * 2),
+         rectangle.append( new Arc2D.Float( x, y + height - (bottomLeftHeight * 2),
                bottomLeftWidth * 2, bottomLeftHeight * 2, 180, 90, Arc2D.OPEN ), true );
       }else{
          rectangle.lineTo( x, y + height );
       }
 
       if( bottomRightWidth > 0 ){
-         rectangle.append( new Arc2D.Double( x + width - (bottomRightWidth * 2), y + height
+         rectangle.append( new Arc2D.Float( x + width - (bottomRightWidth * 2), y + height
                - (bottomRightHeight * 2), bottomRightWidth * 2, bottomRightHeight * 2, 270, 90,
                Arc2D.OPEN ), true );
       }else{
@@ -181,7 +181,7 @@ public class MultiRoundRectangle implements Shape, Cloneable {
       }
 
       if( topRightWidth > 0 ){
-         rectangle.append( new Arc2D.Double( x + width - (topRightWidth * 2), y, topRightWidth * 2,
+         rectangle.append( new Arc2D.Float( x + width - (topRightWidth * 2), y, topRightWidth * 2,
                topRightHeight * 2, 0, 90, Arc2D.OPEN ), true );
       }else{
          rectangle.lineTo( x + width, y );
